+++ loyalty_app/lib/Features/offer_details/presentation/views/widgets/offer_cover_image_and_texts_section.dart
@@ -15,7 +15,7 @@ class OfferCoverImageAndTextsSection extends StatelessWidget {
 
   final String? coverImage;
   final String? offerName;
-  final int? pointAmount;
+  final double? pointAmount;
   final double? userPoints;
 
   @override
